Responding to comments, primarily merging ApiConnector and ContainerInfoFetcher with their child classes
