+++ src/OpenTelemetry.ResourceDetectors.Container/KubernetesContainerInfoFetcher.cs
@@ -1,19 +1,37 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
-
+#if !NETFRAMEWORK
 using System;
+using System.IO;
 using System.Text.Json;
+using System.Text.RegularExpressions;
 
 namespace OpenTelemetry.ResourceDetectors.Container;
 
-internal class KubernetesContainerInfoFetcher : ContainerInfoFetcher
+internal class KubernetesContainerInfoFetcher
 {
+    private static readonly Regex HexStringRegex = new("(^[a-fA-F0-9]+$)", RegexOptions.Compiled);
+
     private readonly string containerName;
 
-    private KubernetesContainerInfoFetcher(ApiConnector? apiConnector, string containerName)
-        : base(apiConnector)
+    private readonly KubeApiConnector? apiConnector;
+
+    private KubernetesContainerInfoFetcher(KubeApiConnector? apiConnector, string containerName)
     {
+        this.apiConnector = apiConnector;
         this.containerName = containerName;
+    }
+
+    public string ExtractContainerId()
+    {
+        // executing request only once. Do we need to retry again if data not available?
+        string response = this.ExecuteApiRequest();
+        if (response == null)
+        {
+            return string.Empty;
+        }
+
+        return this.ParseResponse(response);
     }
 
     internal static KubernetesContainerInfoFetcher? GetInstance()
@@ -25,16 +43,86 @@ internal class KubernetesContainerInfoFetcher : ContainerInfoFetcher
             return new KubernetesContainerInfoFetcher(apiConnector, containerName);
         }
 
-        if (apiConnector != null)
-        {
-            apiConnector.Dispose();
-        }
-
         return null;
     }
 
-    // sample response from kube api given BELOW, which needs to be parsed (ignore whitespace in below sample as real data will not have whitespaces)
-    protected override string ParseResponse(string response)
+    protected static bool CheckAndInitProp(string envPropName1, string? envPropName2, out string? result, bool canContinue = false)
+    {
+        string? value = Environment.GetEnvironmentVariable(envPropName1);
+
+        if (value == null && envPropName2 != null)
+        {
+            value = Environment.GetEnvironmentVariable(envPropName2);
+        }
+
+        if (value == null || string.IsNullOrEmpty(value))
+        {
+            result = null;
+
+            return false;
+        }
+
+        result = value;
+        return true;
+    }
+
+    protected static bool CheckFileAndInitProp(string dirName, string fileName, bool isCertFile, out string result)
+    {
+        result = string.Empty;
+        string filePath = Path.Combine(dirName, fileName);
+        try
+        {
+            FileInfo fileInfo = new(filePath);
+
+            if (!IsFilePresent(fileInfo))
+            {
+                return false;
+            }
+
+            // if this is certificate file, we don't have to read it yet but only check for existence and readability
+            // ca.cert will be directly consumed as input stream when building SSL context
+            if (isCertFile)
+            {
+                result = Path.GetFullPath(filePath);
+            }
+            else
+            {
+                string data = File.ReadAllText(filePath).Trim();
+
+                // file only has whitespaces
+                if (string.IsNullOrEmpty(data))
+                {
+                    return false;
+                }
+
+                result = data;
+            }
+        }
+        catch (Exception e)
+        {
+            ContainerExtensionsEventSource.Log.ExtractResourceAttributesException("Cannot Read " + Path.GetFullPath(filePath) + " : " + e.Message, e);
+            return false;
+        }
+
+        return true;
+    }
+
+    protected static string FormatContainerId(string unFormattedId)
+    {
+        // "containerID"="docker://18e1f4b72f6861b5e591e11ea6db0640377de6ed5dc9bffbae4d9ab284d53044"
+        // Assuming kube api return container id always in this format prefixed with 'docker://'. (Big assumption?)
+        string formattedId = unFormattedId.Substring(unFormattedId.LastIndexOf("/", StringComparison.InvariantCulture) + 1);
+
+        // should be valid hex string
+        if (!HexStringRegex.Match(formattedId).Success)
+        {
+            return string.Empty;
+        }
+
+        return formattedId;
+    }
+
+    protected string ParseResponse(string response)
     {
         // Following https://kubernetes.io/docs/reference/kubernetes-api/workload-resources/pod-v1/#PodStatus
         var obj = JsonSerializer.Deserialize<KubernetesProperties.Pod>(response);
@@ -62,7 +150,12 @@ internal class KubernetesContainerInfoFetcher : ContainerInfoFetcher
         return string.Empty;
     }
 
-    private static bool CheckAndInitRequirements(out ApiConnector? apiConnector, out string? containerName)
+    private static bool IsFilePresent(FileInfo fileInfo)
+    {
+        return fileInfo.Exists && (fileInfo.Length != 0);
+    }
+
+    private static bool CheckAndInitRequirements(out KubeApiConnector? apiConnector, out string? containerName)
     {
         apiConnector = null;
         containerName = null;
@@ -79,7 +172,7 @@ internal class KubernetesContainerInfoFetcher : ContainerInfoFetcher
                 && CheckAndInitProp(KubernetesProperties.HostnameEnvVar, null, out p.HostName)
                 && CheckAndInitProp(
                     KubernetesProperties.ContainerNameEnvVar,
-                    KubernetesProperties.ContainerNameSysProp,
+                    KubernetesProperties.ContainerNameEnvVar2,
                     out p.ContainerName)
 
                 // namespace can be extracted as env or from k8 secret file. More preference to env var. (need to change?)
@@ -115,6 +208,11 @@ internal class KubernetesContainerInfoFetcher : ContainerInfoFetcher
         return false;
     }
 
+    private string ExecuteApiRequest()
+    {
+        return this.apiConnector!.ExecuteRequest();
+    }
+
     private class Properties
     {
         public string? KubernetesHost;
@@ -126,3 +224,4 @@ internal class KubernetesContainerInfoFetcher : ContainerInfoFetcher
         public string? ContainerName;
     }
 }
+#endif
